TURMERIC-1568: fix issue with using eclipse internals during perspective creation.  The JavaPerspectiveFactory and the PageLayout classes were being used directly, when they were marked as internal.
In eclipse 4.1, the PageLayout class was removed entirely from the
org.eclipse.ui.internals package.  Thus when the Turmeric SOA perspective
was opened in eclipse 4.1, the class was not found.

Solution was to copy over the code we needed for creation, of the base
perspective, and then require additional items be added through the
perspectiveExtensions extension point.